migrate Python C extensions to Python 3 API
Update pyextend.c, pydataprocessing.c, and pydatahoneyd.c to use
Python 3 C API instead of Python 2. Changes include:

- Replace PyString_* with PyBytes_* for binary data and PyUnicode_*
  for module names
- Update module initialization from Py_InitModule to PyModule_Create
  with PyModuleDef structure
- Change Py_GetPath/PySys_SetPath to PyRun_SimpleString for sys.path
  manipulation (deprecated functions removed in Python 3)
- Fix all size parameters from int to Py_ssize_t as required by
  Python 3 API (PyBytes_AsStringAndSize, PyMarshalToString)
- Remove unnecessary PyCodeObject* cast from PyEval_EvalCode calls
- Update printf format specifiers from %d to %zd for Py_ssize_t
- Fix escape sequence warning in syslog message

Add conditional compilation in CMakeLists.txt to only build Python
extensions when Python 3 is detected, and link against Python 3
libraries.

Python extensions now compile cleanly with Python 3.13.

+++ pydatahoneyd.c
@@ -181,7 +181,7 @@ pyrecord_test(void)
 	PyObject *pValue, *pRes;
 	struct PyFilter *filter = PyFilterFromCode(some_code);
 	char *result;
-	int res_len;
+	Py_ssize_t res_len;
 	struct record *record = NULL;
 	struct evbuffer *tmp = evbuffer_new();
 	assert(filter != NULL);
@@ -218,8 +218,8 @@ pyrecord_test(void)
 
 		Py_DECREF(pRes);
 		Py_DECREF(pValue);
-	
-		fprintf(stderr, "\t\tResult len: %d\n", res_len);
+
+		fprintf(stderr, "\t\tResult len: %zd\n", res_len);
 	
 		free(record);
 	}
